config and table tests moved to tape
